third_party/blink/web_tests/platform/generic/fast/frames/sandboxed-iframe-attribute-parsing-14-expected.txt

CONSOLE ERROR: Blocked script execution in 'sandboxed-iframe-attribute-parsing-disallowed.html' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
Full sandbox.

PASS
